refactor(jazz-tools): Phase 3 (opfs_btree) — split cfg-gated constructors into native/wasm submodules
storage/opfs_btree.rs is renamed to storage/opfs_btree/mod.rs and the
cfg-gated entry points move to dedicated, module-level cfg-gated
submodules:

  storage/opfs_btree/
    mod.rs    — shared body (AnyFile dispatch enum, SyncFile impl on
                AnyFile, OpfsBTreeStorage struct, the Storage trait
                impl, helpers, tests)
    native.rs — #[cfg(not(target_arch = "wasm32"))]; OpfsBTreeStorage::open
                (taking impl AsRef<Path>) backed by opfs_btree::StdFile
    wasm.rs   — #[cfg(target_arch = "wasm32")]; OpfsBTreeStorage::with_opfs,
                ::open_opfs, ::destroy_opfs backed by opfs_btree::OpfsFile

The shared `AnyFile` enum still has cfg-gated `Std` / `Opfs` variants and
the SyncFile dispatch arms still match on those — that's the natural
shape of a typed dispatch enum and would only churn if every native- or
wasm-specific code path were duplicated. The substantive win is removing
all the cfg-gated *function bodies* threaded through the shared file:
everything in `native.rs` is unconditionally native, everything in
`wasm.rs` is unconditionally wasm.

Visibility surgery: `AnyFile` becomes `pub(super)` and `open_with_file`
+ `map_storage_err` become `pub(super) fn` so the new submodules can
delegate to the shared core.

Spec: specs/todo/c_later/crate_structure_cleanup.md (Phase 3 — opfs_btree
cfg-fork half; the storage trait/impl split landed earlier in PR #711).
